:sparkles: Add new Transform method to Node
When there are several converts that applies from and to the same
types, the PluginManager is not able to detect which one is the correct.
In that case, the user needs to specify which converter to use. For instance
this happen with transformation from Binary to NodeContainer formats.
This new method simplify and avoid the creation of the converter by the user.

Closes #17