:bug: Dispose children nodes when removing or disposing parent
If the parent node is disposed the children shall be disposed too.
Imagine we are disposing the pack format, it makes no sense to keep
the children since the parent doesn't exist anymore and it can cause
problem accessing to the parent back. The same logic applies when
removing the children from parent.

Fixes #20